requirements.txt
django
python-dotenv

#ses
boto3

#testing
pytest-django
